RFC-004 P5: handler wave 2 — browser_web, coding, devops domains
13 handlers move off ToolExecutor:
- browser_web.py: 5 browser handlers + _browser_with_bulkhead,
  web_search, fetch_url, http_probe (browser + static web share one
  module per plan advisory #8)
- coding.py: claude_code + _parse_claude_stream_json (sole consumer)
- devops.py: git_ops, kubectl, docker_ops, terraform_ops

HandlerDeps grows two live accessors (browser_manager, bulkheads);
same late-resolution seam as wave 1. Declared mechanical adjustments:
lazy relative imports re-anchored one level; nothing else changed
(AST-comparable).

Table rebinds 13 entries to the new owners; public attrs
browser_web_tools / coding_tools / devops_tools follow the P4
convention. Direct-call test sites re-pointed (terraform, docker,
http_probe, bulkhead); the two remaining __new__-bypass fixtures now
construct properly (live deps keep their overrides governing); 5
now-unused result assignments dropped; orphaned shlex import removed.

executor.py 1,530 -> 1,298 lines. Full suite 6,125 passed; lint gate
new=0 (8 more pre-existing resolved).
